CLDR-16390 move components into separate .js file
- also add a-timeline*
