+++ tools/cldr-apps/js/src/cldrVueRouter.js
@@ -1,34 +1,9 @@
-import { specialToComponent } from "./specialToComponentMap";
-
+// Note: the order of imports is important here. Possibly a race condition in globals inside some of the .js files.
 import { createApp } from "vue";
-
+import {  notification } from "ant-design-vue";
+import { specialToComponent } from "./specialToComponentMap";
 import { getCldrOpts } from "./getCldrOpts";
-
-// See index.js for css imports.
-// example: import {SomeComponent} from 'whatever'
-import {
-  // components
-  Alert,
-  Button,
-  Checkbox,
-  Form,
-  Input,
-  List,
-  Popover,
-  Progress,
-  Radio,
-  Spin,
-  Steps,
-  Tooltip,
-  // functions
-  notification,
-} from "ant-design-vue";
-
-// local components
-import CldrValue from "./views/CldrValue.vue";
-import LoginButton from "./views/LoginButton.vue";
-import ReportResponse from "./views/ReportResponse.vue";
-import SearchButton from "./views/SearchButton.vue";
+import { setupComponents} from './setupComponents';
 
 /**
  * The App created and mounted most recently. For .unmount …
@@ -123,37 +98,4 @@ function errBox(e, operation) {
   });
 }
 
-/**
- * add any Vue components needed here.
- * @param {App} app
- */
-function setupComponents(app) {
-  // example:
-  // app.component('SomeComponent', SomeComponent)
-  // Keep this list sorted
-  app.component("a-alert", Alert);
-  app.component("a-button", Button);
-  app.component("a-checkbox", Checkbox);
-  app.component("a-form-item", Form.Item);
-  app.component("a-form", Form);
-  app.component("a-input-password", Input.Password);
-  app.component("a-input-search", Input.Search);
-  app.component("a-input", Input);
-  app.component("a-list", List);
-  app.component("a-list-item", List.Item);
-  app.component("a-list-item-meta", List.Item.Meta);
-  app.component("a-popover", Popover);
-  app.component("a-progress", Progress);
-  app.component("a-radio-group", Radio.Group);
-  app.component("a-radio", Radio);
-  app.component("a-spin", Spin);
-  app.component("a-step", Steps.Step);
-  app.component("a-steps", Steps);
-  app.component("a-tooltip", Tooltip);
-  app.component("cldr-loginbutton", LoginButton);
-  app.component("cldr-report-response", ReportResponse);
-  app.component("cldr-searchbutton", SearchButton);
-  app.component("cldr-value", CldrValue);
-}
-
-export { showPanel, createCldrApp, setupComponents };
+export { showPanel, createCldrApp };
